requirements.txt
opencv-python
mediapipe
numpy
Pillow
openai
PyQt6
python-dotenv
reportlab
pytest
requests
